fix: allow for large reltuples for platform safety

+++ pkg/pg/catalog/pg_class.go
@@ -66,7 +66,7 @@ type PgClassRow struct {
 	SchemaName string  `json:"schemaname"`
 	RelName    string  `json:"relname"`
 	RelTuples  float64 `json:"reltuples"`
-	RelPages   int     `json:"relpages"`
+	RelPages   int64   `json:"relpages"`
 }
 
 // PgClassPayload is the JSON body POSTed to /api/v1/agent/pg_class.
